add delete account

+++ src/main/java/com/projectjava/UserManager.java
@@ -6,20 +6,27 @@ import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.result.DeleteResult;
 
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.security.Keys;
+
+import java.security.Key;
+import java.util.Date;
+
 import org.bson.Document;
 
 public class UserManager {
 
-    private static final MongoClient mongoClient;
-
-    static {
-        mongoClient = MongoClients.create("mongodb+srv://Guillaume:test@cluster0.pfjuogp.mongodb.net/javaserver");
-    }
-
     private final MongoCollection<Document> userCollection;
+    private final Key jwtKey;
 
-    public UserManager() {
+    public UserManager(Key jwtKey) {
+        this.jwtKey = jwtKey;
+
+        MongoClient mongoClient = MongoClients
+                .create("mongodb+srv://Guillaume:test@cluster0.pfjuogp.mongodb.net/javaserver");
         MongoDatabase database = mongoClient.getDatabase("javaserver");
+
         userCollection = database.getCollection("users");
     }
 
@@ -37,6 +44,26 @@ public class UserManager {
         }
     }
 
+    public boolean logoutUser(String username, String authToken) {
+        if (isValidAuthToken(username, authToken)) {
+            System.out.println("User logged out successfully!");
+            return true;
+        } else {
+            System.err.println("Invalid authentication token. Logout failed.");
+            return false;
+        }
+    }
+
+    public String generateAuthToken(String username) {
+        Date expiration = new Date(System.currentTimeMillis() + 3600000); // Expiration dans 1 heure
+
+        return Jwts.builder()
+                .setSubject(username)
+                .setExpiration(expiration)
+                .signWith(jwtKey)
+                .compact();
+    }
+
     public boolean authenticateUser(String username, String password) {
         Document query = new Document("username", username).append("password", password);
 
@@ -45,18 +72,25 @@ public class UserManager {
         return count == 1;
     }
 
-    public boolean deleteUser(String username) {
-        Document query = new Document("username", username);
-
-        try {
-            DeleteResult result = userCollection.deleteOne(query);
-            System.out.println("Deleted " + result.getDeletedCount() + " user(s) successfully.");
-            return result.getDeletedCount() == 1;
-        } catch (Exception e) {
-            System.err.println("Failed to delete user. Error: " + e.getMessage());
-            e.printStackTrace();
+    public boolean deleteUser(String username, String authToken) {
+        if (isValidAuthToken(username, authToken)) {
+            Document query = new Document("username", username);
+            userCollection.deleteOne(query);
+            System.out.println("User deleted successfully!");
+            return true;
+        } else {
+            System.err.println("Invalid authentication token. User not deleted.");
             return false;
         }
     }
 
+    private boolean isValidAuthToken(String username, String authToken) {
+        try {
+            Jwts.parserBuilder().setSigningKey(jwtKey).build().parseClaimsJws(authToken);
+            return true;
+        } catch (Exception e) {
+            System.err.println("Invalid authentication token. Error: " + e.getMessage());
+            return false;
+        }
+    }
 }
